Refactored the implementation of theory macros in internal/
